+++ attr/Cargo.toml
@@ -11,7 +11,7 @@ rust-version = { workspace = true }
 proc-macro = true
 
 [dependencies]
+heck = { workspace = true }
 proc-macro2 = { workspace = true }
-generic-array-struct-common = { workspace = true }
 quote = { workspace = true }
-syn = { workspace = true, features = ["printing", "parsing", "proc-macro"]}
+syn = { workspace = true, features = ["derive", "printing", "parsing", "proc-macro"]}
