+++ attr/src/errs.rs
@@ -13,17 +13,17 @@ const ONLY_WORKS_WITH_STRUCTS_WITH_NAMED_FIELDS_ERRMSG: &str =
 
 /// Panic with `err` error message
 #[inline]
-pub fn proc_macro_error(err: &str) -> ! {
+pub(crate) fn proc_macro_error(err: &str) -> ! {
     panic!("{MACRO_NAME} {err}")
 }
 
 #[inline]
-pub fn panic_req_single_generic() -> ! {
+pub(crate) fn panic_req_single_generic() -> ! {
     proc_macro_error(REQ_SINGLE_GENERIC_TYPE_PARAM_ERRMSG)
 }
 
 #[inline]
-pub fn panic_req_all_fields_same_generic() -> ! {
+pub(crate) fn panic_req_all_fields_same_generic() -> ! {
     proc_macro_error(REQ_ALL_FIELDS_SAME_GENERIC_TYPE_ERRMSG)
 }
