--- common/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "generic-array-struct-common"
-version = { workspace = true }
-edition = { workspace = true }
-license-file = { workspace = true }
-description = { workspace = true }
-repository = { workspace = true }
-
-[dependencies]
-heck = { workspace = true }
-quote = { workspace = true }
-syn = { workspace = true, features = ["derive"] }
